add Max's sigchain, to test revoking eldest link
We're adding this to cover a recent break in `keybase id max`, when we
tried to make sure eldest key delegation info was initialized
consistently. That break was reverted in master, and this test will make
sure that we don't break it again the second time.

Note that we're making another tweak in client at the same time
(IsDirectlyRevoked), to clean up the way we count revoked links in the
sigchain tests. The unrevoked length number we're checking with the new
test case here assumes that tweak.